Restructured functions and renamed file
Added gitignore
